refactor: Move page contents to components/pages
